use size_t for size-related fields instead of int

+++ include/nfa.h
@@ -1,5 +1,5 @@
-#include "dynarr.h"
 #include "bitmask.h"
+#include "dynarr.h"
 #include "matcher.h"
 
 #ifndef NFA_H
@@ -9,28 +9,27 @@
 
 typedef struct transition {
     matcher_t matcher;
-    int to_state;
+    size_t to_state;
 } transition_t;
 
 /* Thompson epsilon-NFA: only one start and one finish.
    edge stored as linked-lists */
 typedef struct tepsnfa {
-    int state_num;
-    int start_state;
-    int final_state;
+    size_t state_num;
+    size_t start_state;
+    size_t final_state;
     /* type: dynarr of transtion. index is starting state. */
     dynarr_t state_transitions;
 } tepsnfa;
 
 /* epsilon NFA: edges stored as an adjacent matrix */
 typedef struct epsnfa {
-    int state_num;
+    size_t state_num;
     matcher_t* transition_table;
     bitmask_t is_start;
     bitmask_t is_finish;
     dynarr_t char_class_pool; /* type: char_class_t */
 } epsnfa;
-
 
 /* Thompson epsilon-NFA methods */
 
@@ -38,13 +37,15 @@ typedef struct epsnfa {
    transition between them */
 tepsnfa tepsnfa_one_transition(matcher_t m);
 
-int tepsnfa_print(tepsnfa* self);
+size_t tepsnfa_print(tepsnfa* self);
 
 void tepsnfa_clear(tepsnfa* self);
 
 void tepsnfa_add_state(tepsnfa* self);
 
-void tepsnfa_add_transition(tepsnfa* self, int from, matcher_t matcher, int to);
+void tepsnfa_add_transition(
+    tepsnfa* self, size_t from, matcher_t matcher, size_t to
+);
 
 tepsnfa tepsnfa_deepcopy(const tepsnfa* self);
 
@@ -65,14 +66,13 @@ void tepsnfa_to_opt(tepsnfa* self);
 
 epsnfa tepsnfa_to_epsnfa_and_reduce_eps(tepsnfa* input);
 
-
 /* Epsilon-NFA methods */
 
-epsnfa epsnfa_new(const int state_num);
+epsnfa epsnfa_new(const size_t state_num);
 
-void epsnfa_get_eps_closure(epsnfa* self, int state, bitmask_t* closure);
+void epsnfa_get_eps_closure(epsnfa* self, size_t state, bitmask_t* closure);
 
-void epsnfa_reduce_eps(epsnfa* self, int state);
+void epsnfa_reduce_eps(epsnfa* self, size_t state);
 
 void epsnfa_print(epsnfa* self);
 
@@ -94,13 +94,11 @@ size_t epsnfa_find_initial_match(
     const size_t start_offset
 );
 
-dynarr_t
-epsnfa_find_matches(
+dynarr_t epsnfa_find_matches(
     const epsnfa* epsnfa, const char* input, const int is_global
 );
 
-dynarr_t
-epsnfa_find_matches_multiline(
+dynarr_t epsnfa_find_matches_multiline(
     const epsnfa* epsnfa, const char* input, const int is_global
 );
